requirements.txt
edge_tts==6.1.12
pygame==2.6.0
sounddevice==0.5.0
vosk==0.3.45
google-generativeai==0.7.2
python-dotenv
